refactor: split commands.rs into focused sub-modules (#671)
Split the monolithic commands.rs (3021 lines) into a commands/ directory
with 7 focused modules:

- mod.rs: route_command dispatcher and CommandResult enum
- handlers.rs: /who, /set_status, /subscribe, /unsubscribe handlers
- info.rs: /info and /help command handlers
- plugin.rs: dispatch_plugin, CrossRoomContext, cross-room dispatch
- team.rs: /team command handler
- validate.rs: parameter validation against plugin schemas
- tests.rs: all existing unit tests (preserved from original file)

Pure refactor — no behavioral changes. All 1597 tests pass.
Incorporates CrossRoomContext from #681 (PR #711).

Closes #671
