+++ crates/room-daemon/src/broker/commands/tests.rs
@@ -1,1043 +1,7 @@
-use std::panic::AssertUnwindSafe;
-use std::time::Duration;
-
-use futures_util::FutureExt;
-use room_protocol::{EventFilter, SubscriptionTier};
-
-use room_protocol::{make_system, Message};
-
-use crate::plugin::{
-    builtin_command_infos, snapshot_metadata, ChatWriter, CommandContext, CommandInfo,
-    HistoryReader, ParamType, PluginResult,
-};
-
-use super::{
-    admin::{handle_admin_cmd, ADMIN_CMD_NAMES},
-    auth::check_join_permission,
-    fanout::broadcast_and_persist,
-    persistence::{persist_event_filters, persist_subscriptions},
-    state::RoomState,
-};
-
-/// Maximum time a plugin is allowed to run before being cancelled.
-const PLUGIN_TIMEOUT: Duration = Duration::from_secs(30);
-
-/// The result of routing an inbound command line.
-pub(crate) enum CommandResult {
-    /// The command was fully handled with a broadcast or no-op; nothing to send back privately.
-    Handled,
-    /// The command was handled with a broadcast; oneshot callers should receive this JSON echo.
-    ///
-    /// Interactive clients already receive the message via the broadcast channel, so
-    /// `handle_client` treats this identically to `Handled`. One-shot senders are not
-    /// subscribed to the broadcast, so `handle_oneshot_send` writes the JSON back to them
-    /// directly, avoiding the EOF parse error the client would otherwise see.
-    HandledWithReply(String),
-    /// The command was handled privately; send this JSON line back only to the issuer.
-    Reply(String),
-    /// The command was handled and the broker is shutting down.
-    Shutdown,
-    /// Not a special command; the caller should broadcast or DM `msg` normally.
-    Passthrough(Message),
-}
-
-/// Route a parsed `Message` for a given `username` against `state`.
-///
-/// Thin dispatcher that matches on command name and delegates to dedicated
-/// `handle_*` functions. For non-command messages (regular chat, DMs) returns
-/// `CommandResult::Passthrough(msg)` so the caller can broadcast or DM it.
-///
-/// This is the unified entry point used by both the interactive inbound task
-/// and `handle_oneshot_send`.
-pub(crate) async fn route_command(
-    msg: Message,
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    if let Message::Command {
-        ref cmd,
-        ref params,
-        ..
-    } = msg
-    {
-        // Validate params against built-in schema before dispatching.
-        let builtins = builtin_command_infos();
-        if let Some(cmd_info) = builtins.iter().find(|c| c.name == *cmd) {
-            if let Err(err_msg) = validate_params(params, cmd_info) {
-                let sys = make_system(&state.room_id, "broker", err_msg);
-                let json = serde_json::to_string(&sys)?;
-                return Ok(CommandResult::Reply(json));
-            }
-        }
-
-        match cmd.as_str() {
-            "who" => return handle_who(state).await,
-            "who_all" => return handle_who_all(state).await,
-            "set_status" => return handle_set_status(params, username, state).await,
-            "subscribe" | "set_subscription" => {
-                return handle_subscribe(params, username, state).await
-            }
-            "unsubscribe" => return handle_unsubscribe(username, state).await,
-            "subscribe_events" | "set_event_filter" => {
-                return handle_subscribe_events(params, username, state).await
-            }
-            "subscriptions" => return handle_subscriptions(state).await,
-            "info" | "room-info" => return handle_info_cmd(params, state).await,
-            "help" => return handle_help_cmd(params, state),
-            "team" => return handle_team(params, username, state).await,
-            _ if ADMIN_CMD_NAMES.contains(&cmd.as_str()) => {
-                return handle_admin(cmd, params, username, state).await
-            }
-            _ => {}
-        }
-
-        // Plugin dispatch — check registry before falling through to Passthrough.
-        if let Some(plugin) = state.plugin_registry.resolve(cmd) {
-            let plugin_name = plugin.name().to_owned();
-            match dispatch_plugin(plugin, &msg, username, state).await {
-                Ok(result) => return Ok(result),
-                Err(e) => {
-                    let err_msg = format!("plugin:{plugin_name} error: {e}");
-                    let sys = make_system(&state.room_id, "broker", err_msg);
-                    let json = serde_json::to_string(&sys)?;
-                    return Ok(CommandResult::Reply(json));
-                }
-            }
-        }
-    }
-
-    Ok(CommandResult::Passthrough(msg))
-}
-
-// ── Command handlers ─────────────────────────────────────────────────────────
-
-/// Handle `/who` — list online users with their statuses.
-async fn handle_who(state: &RoomState) -> anyhow::Result<CommandResult> {
-    let entries: Vec<String> = state
-        .status_entries()
-        .await
-        .into_iter()
-        .map(|(u, s)| {
-            if s.is_empty() {
-                u
-            } else {
-                // Sanitize commas in status text so the TUI parser
-                // (which splits on ", ") doesn't treat status fragments
-                // as separate usernames (#656).
-                let safe = s.replace(", ", "; ");
-                format!("{u}: {safe}")
-            }
-        })
-        .collect();
-    let content = if entries.is_empty() {
-        "no users online".to_owned()
-    } else {
-        format!("online — {}", entries.join(", "))
-    };
-    let sys = make_system(&state.room_id, "broker", content);
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::Reply(json))
-}
-
-/// Handle `/who_all` — list all daemon-registered users (cross-room).
-///
-/// In daemon mode, queries the [`UserRegistry`] for every registered username.
-/// In standalone mode, falls back to the same output as `/who`.
-async fn handle_who_all(state: &RoomState) -> anyhow::Result<CommandResult> {
-    let usernames: Vec<String> = if let Some(registry) = state.auth.registry.get() {
-        let guard = registry.lock().await;
-        let mut names: Vec<String> = guard
-            .list_users()
-            .iter()
-            .map(|u| u.username.clone())
-            .collect();
-        names.sort();
-        names
-    } else {
-        // Standalone mode — fall back to current room users.
-        state
-            .status_entries()
-            .await
-            .into_iter()
-            .map(|(u, _)| u)
-            .collect()
-    };
-    let json_array = serde_json::to_string(&usernames)?;
-    let content = format!("users_all: {json_array}");
-    let sys = make_system(&state.room_id, "broker", content);
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::Reply(json))
-}
-
-/// Handle `/set_status [text]` — set or clear the user's status, broadcast the change.
-async fn handle_set_status(
-    params: &[String],
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    let status = params.join(" ");
-    state.set_status(username, status.clone()).await;
-    let display = if status.is_empty() {
-        format!("{username} cleared their status")
-    } else {
-        format!("{username} set status: {status}")
-    };
-    let sys = make_system(&state.room_id, "broker", display);
-    let seq_msg =
-        broadcast_and_persist(&sys, &state.clients, &state.chat_path, &state.seq_counter).await?;
-    let json = serde_json::to_string(&seq_msg)?;
-    Ok(CommandResult::HandledWithReply(json))
-}
-
-/// Handle `/subscribe [tier]` — subscribe to room messages with a tier filter.
-async fn handle_subscribe(
-    params: &[String],
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    // Reject subscriptions from users who cannot join this room.
-    // Without this guard, non-participants could subscribe to DM rooms
-    // and read regular messages through the poll path (#491).
-    if let Err(reason) = check_join_permission(username, state.config.as_ref()) {
-        let sys = make_system(&state.room_id, "broker", reason);
-        let json = serde_json::to_string(&sys)?;
-        return Ok(CommandResult::Reply(json));
-    }
-    let tier_str = params.first().map(String::as_str).unwrap_or("full");
-    let tier: SubscriptionTier = match tier_str.parse() {
-        Ok(t) => t,
-        Err(e) => {
-            let sys = make_system(&state.room_id, "broker", e);
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    };
-    state.set_subscription(username, tier).await;
-    persist_subscriptions(state).await;
-    let display = format!("{username} subscribed to {} (tier: {tier})", state.room_id);
-    let sys = make_system(&state.room_id, "broker", display);
-    broadcast_and_persist(&sys, &state.clients, &state.chat_path, &state.seq_counter).await?;
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::HandledWithReply(json))
-}
-
-/// Handle `/unsubscribe` — set subscription to Unsubscribed tier.
-async fn handle_unsubscribe(username: &str, state: &RoomState) -> anyhow::Result<CommandResult> {
-    state
-        .set_subscription(username, SubscriptionTier::Unsubscribed)
-        .await;
-    persist_subscriptions(state).await;
-    let display = format!("{username} unsubscribed from {}", state.room_id);
-    let sys = make_system(&state.room_id, "broker", display);
-    broadcast_and_persist(&sys, &state.clients, &state.chat_path, &state.seq_counter).await?;
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::HandledWithReply(json))
-}
-
-/// Handle `/subscribe_events [filter]` — set event filter for the user.
-async fn handle_subscribe_events(
-    params: &[String],
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    // Reject from users who cannot join this room (same guard as subscribe).
-    if let Err(reason) = check_join_permission(username, state.config.as_ref()) {
-        let sys = make_system(&state.room_id, "broker", reason);
-        let json = serde_json::to_string(&sys)?;
-        return Ok(CommandResult::Reply(json));
-    }
-    let filter_str = params.first().map(String::as_str).unwrap_or("all");
-    let filter: EventFilter = match filter_str.parse() {
-        Ok(f) => f,
-        Err(e) => {
-            let sys = make_system(&state.room_id, "broker", e);
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    };
-    state.set_event_filter(username, filter.clone()).await;
-    persist_event_filters(state).await;
-    let display = format!(
-        "{username} event filter set to {filter} in {}",
-        state.room_id
-    );
-    let sys = make_system(&state.room_id, "broker", display);
-    broadcast_and_persist(&sys, &state.clients, &state.chat_path, &state.seq_counter).await?;
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::HandledWithReply(json))
-}
-
-/// Handle `/subscriptions` — list all subscription tiers and event filters.
-async fn handle_subscriptions(state: &RoomState) -> anyhow::Result<CommandResult> {
-    let raw = state.subscription_entries().await;
-    let ef_raw = state.event_filter_entries().await;
-    let content = if raw.is_empty() && ef_raw.is_empty() {
-        "no subscriptions".to_owned()
-    } else {
-        let mut parts = Vec::new();
-        if !raw.is_empty() {
-            let entries: Vec<String> = raw.into_iter().map(|(u, t)| format!("{u}: {t}")).collect();
-            parts.push(format!("tiers — {}", entries.join(", ")));
-        }
-        if !ef_raw.is_empty() {
-            let entries: Vec<String> = ef_raw
-                .into_iter()
-                .map(|(u, f)| format!("{u}: {f}"))
-                .collect();
-            parts.push(format!("event filters — {}", entries.join(", ")));
-        }
-        parts.join(" | ")
-    };
-    let sys = make_system(&state.room_id, "broker", content);
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::Reply(json))
-}
-
-/// Handle `/info [username]` — dispatch to room info or user info.
-async fn handle_info_cmd(params: &[String], state: &RoomState) -> anyhow::Result<CommandResult> {
-    let result = handle_info(params, state).await;
-    let sys = make_system(&state.room_id, "broker", result);
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::Reply(json))
-}
-
-/// Handle `/help [command]` — list commands or show detail for one.
-fn handle_help_cmd(params: &[String], state: &RoomState) -> anyhow::Result<CommandResult> {
-    let reply = handle_help(params, state);
-    let sys = make_system(&state.room_id, "broker", reply);
-    let json = serde_json::to_string(&sys)?;
-    Ok(CommandResult::Reply(json))
-}
-
-/// Handle `/team <action> [args...]` — manage daemon-level teams.
-///
-/// Subcommands:
-/// - `/team join <team> [user]` — add yourself or named user to a team (creates if needed)
-/// - `/team leave <team> [user]` — remove yourself or named user (deletes if empty)
-/// - `/team list` — show all teams and members
-/// - `/team show <team>` — show members of a specific team
-async fn handle_team(
-    params: &[String],
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    let action = params.first().map(String::as_str).unwrap_or("");
-    let registry_lock = match state.auth.registry.get() {
-        Some(r) => r,
-        None => {
-            let sys = make_system(
-                &state.room_id,
-                "broker",
-                "teams require daemon mode".to_owned(),
-            );
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    };
-
-    match action {
-        "join" => {
-            let team_name = match params.get(1) {
-                Some(t) => t.as_str(),
-                None => {
-                    let sys = make_system(
-                        &state.room_id,
-                        "broker",
-                        "usage: /team join <team> [user]".to_owned(),
-                    );
-                    let json = serde_json::to_string(&sys)?;
-                    return Ok(CommandResult::Reply(json));
-                }
-            };
-            let target_user = params.get(2).map(|s| s.as_str()).unwrap_or(username);
-            let mut reg = registry_lock.lock().await;
-            match reg.join_team(team_name, target_user) {
-                Ok(true) => {
-                    drop(reg);
-                    let content = format!("{target_user} joined team {team_name}");
-                    let sys = make_system(&state.room_id, "broker", content);
-                    let seq_msg = broadcast_and_persist(
-                        &sys,
-                        &state.clients,
-                        &state.chat_path,
-                        &state.seq_counter,
-                    )
-                    .await?;
-                    let json = serde_json::to_string(&seq_msg)?;
-                    Ok(CommandResult::HandledWithReply(json))
-                }
-                Ok(false) => {
-                    let sys = make_system(
-                        &state.room_id,
-                        "broker",
-                        format!("{target_user} is already in team {team_name}"),
-                    );
-                    let json = serde_json::to_string(&sys)?;
-                    Ok(CommandResult::Reply(json))
-                }
-                Err(e) => {
-                    let sys = make_system(&state.room_id, "broker", format!("error: {e}"));
-                    let json = serde_json::to_string(&sys)?;
-                    Ok(CommandResult::Reply(json))
-                }
-            }
-        }
-        "leave" => {
-            let team_name = match params.get(1) {
-                Some(t) => t.as_str(),
-                None => {
-                    let sys = make_system(
-                        &state.room_id,
-                        "broker",
-                        "usage: /team leave <team> [user]".to_owned(),
-                    );
-                    let json = serde_json::to_string(&sys)?;
-                    return Ok(CommandResult::Reply(json));
-                }
-            };
-            let target_user = params.get(2).map(|s| s.as_str()).unwrap_or(username);
-            let mut reg = registry_lock.lock().await;
-            match reg.leave_team(team_name, target_user) {
-                Ok(true) => {
-                    drop(reg);
-                    let content = format!("{target_user} left team {team_name}");
-                    let sys = make_system(&state.room_id, "broker", content);
-                    let seq_msg = broadcast_and_persist(
-                        &sys,
-                        &state.clients,
-                        &state.chat_path,
-                        &state.seq_counter,
-                    )
-                    .await?;
-                    let json = serde_json::to_string(&seq_msg)?;
-                    Ok(CommandResult::HandledWithReply(json))
-                }
-                Ok(false) => {
-                    let sys = make_system(
-                        &state.room_id,
-                        "broker",
-                        format!("{target_user} is not in team {team_name}"),
-                    );
-                    let json = serde_json::to_string(&sys)?;
-                    Ok(CommandResult::Reply(json))
-                }
-                Err(e) => {
-                    let sys = make_system(&state.room_id, "broker", format!("error: {e}"));
-                    let json = serde_json::to_string(&sys)?;
-                    Ok(CommandResult::Reply(json))
-                }
-            }
-        }
-        "list" => {
-            let reg = registry_lock.lock().await;
-            let teams = reg.list_teams();
-            let content = if teams.is_empty() {
-                "no teams".to_owned()
-            } else {
-                teams
-                    .iter()
-                    .map(|t| {
-                        let members: Vec<&str> = t.members.iter().map(String::as_str).collect();
-                        format!("{}: {}", t.name, members.join(", "))
-                    })
-                    .collect::<Vec<_>>()
-                    .join("\n")
-            };
-            let sys = make_system(&state.room_id, "broker", content);
-            let json = serde_json::to_string(&sys)?;
-            Ok(CommandResult::Reply(json))
-        }
-        "show" => {
-            let team_name = match params.get(1) {
-                Some(t) => t.as_str(),
-                None => {
-                    let sys = make_system(
-                        &state.room_id,
-                        "broker",
-                        "usage: /team show <team>".to_owned(),
-                    );
-                    let json = serde_json::to_string(&sys)?;
-                    return Ok(CommandResult::Reply(json));
-                }
-            };
-            let reg = registry_lock.lock().await;
-            let content = match reg.get_team(team_name) {
-                Some(team) => {
-                    let members: Vec<&str> = team.members.iter().map(String::as_str).collect();
-                    format!("team {}: {}", team.name, members.join(", "))
-                }
-                None => format!("team not found: {team_name}"),
-            };
-            let sys = make_system(&state.room_id, "broker", content);
-            let json = serde_json::to_string(&sys)?;
-            Ok(CommandResult::Reply(json))
-        }
-        "" => {
-            let sys = make_system(
-                &state.room_id,
-                "broker",
-                "usage: /team <join|leave|list|show> [args...]".to_owned(),
-            );
-            let json = serde_json::to_string(&sys)?;
-            Ok(CommandResult::Reply(json))
-        }
-        other => {
-            let sys = make_system(
-                &state.room_id,
-                "broker",
-                format!("unknown action: {other}. use: join, leave, list, show"),
-            );
-            let json = serde_json::to_string(&sys)?;
-            Ok(CommandResult::Reply(json))
-        }
-    }
-}
-
-/// Handle admin commands (`/kick`, `/reauth`, `/clear-tokens`, `/exit`, `/clear`).
-async fn handle_admin(
-    cmd: &str,
-    params: &[String],
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    let cmd_line = format!("{cmd} {}", params.join(" "));
-    let error = handle_admin_cmd(&cmd_line, username, state).await;
-    if let Some(err) = error {
-        let sys = make_system(&state.room_id, "broker", err);
-        let json = serde_json::to_string(&sys)?;
-        return Ok(CommandResult::Reply(json));
-    }
-    if cmd == "exit" {
-        return Ok(CommandResult::Shutdown);
-    }
-    Ok(CommandResult::Handled)
-}
-
-/// Validate `params` against a command's [`CommandInfo`] schema.
-///
-/// Returns `Ok(())` if all constraints pass, or `Err(message)` with a
-/// human-readable error suitable for sending back as a reply.
-///
-/// Validation rules:
-/// - Required params must be present (not blank).
-/// - `ParamType::Choice` values must be in the allowed set.
-/// - `ParamType::Number` values must parse as `i64` and respect min/max bounds.
-/// - `ParamType::Text` and `ParamType::Username` are accepted as-is (no
-///   server-side validation — username existence is not checked here).
-fn validate_params(params: &[String], schema: &CommandInfo) -> Result<(), String> {
-    for (i, ps) in schema.params.iter().enumerate() {
-        let value = params.get(i).map(String::as_str).unwrap_or("");
-        if ps.required && value.is_empty() {
-            return Err(format!(
-                "/{}: missing required parameter <{}>",
-                schema.name, ps.name
-            ));
-        }
-        if value.is_empty() {
-            continue;
-        }
-        match &ps.param_type {
-            ParamType::Choice(allowed) => {
-                if !allowed.iter().any(|a| a == value) {
-                    return Err(format!(
-                        "/{}: <{}> must be one of: {}",
-                        schema.name,
-                        ps.name,
-                        allowed.join(", ")
-                    ));
-                }
-            }
-            ParamType::Number { min, max } => {
-                let Ok(n) = value.parse::<i64>() else {
-                    return Err(format!(
-                        "/{}: <{}> must be a number, got '{}'",
-                        schema.name, ps.name, value
-                    ));
-                };
-                if let Some(lo) = min {
-                    if n < *lo {
-                        return Err(format!("/{}: <{}> must be >= {lo}", schema.name, ps.name));
-                    }
-                }
-                if let Some(hi) = max {
-                    if n > *hi {
-                        return Err(format!("/{}: <{}> must be <= {hi}", schema.name, ps.name));
-                    }
-                }
-            }
-            ParamType::Text | ParamType::Username => {}
-        }
-    }
-    Ok(())
-}
-
-/// Extract `--room <room_id>` from a params list, returning the target room ID
-/// and the cleaned params with the flag and its value stripped out.
-///
-/// Returns `None` if no `--room` flag is present. The flag can appear at any
-/// position in the params list; both it and the subsequent value are removed
-/// so the plugin never sees them.
-fn extract_room_flag(params: &[String]) -> Option<(String, Vec<String>)> {
-    let mut target_room = None;
-    let mut cleaned = Vec::with_capacity(params.len());
-    let mut skip_next = false;
-
-    for (i, p) in params.iter().enumerate() {
-        if skip_next {
-            skip_next = false;
-            continue;
-        }
-        if p == "--room" {
-            if let Some(val) = params.get(i + 1) {
-                target_room = Some(val.clone());
-                skip_next = true;
-                continue;
-            }
-        }
-        cleaned.push(p.clone());
-    }
-
-    target_room.map(|room_id| (room_id, cleaned))
-}
-
-/// Build a [`CommandContext`] from the given command fields and room state.
-///
-/// This is the single place where `CommandContext` is assembled — both
-/// `dispatch_plugin` (local) and `dispatch_cross_room` (remote) call it.
-async fn build_command_context(
-    cmd: &str,
-    params: &[String],
-    id: &str,
-    ts: chrono::DateTime<chrono::Utc>,
-    username: &str,
-    state: &RoomState,
-    plugin_name: &str,
-) -> CommandContext {
-    let history = HistoryReader::new(&state.chat_path, username);
-    let writer = ChatWriter::new(
-        &state.clients,
-        &state.chat_path,
-        &state.room_id,
-        &state.seq_counter,
-        plugin_name,
-    );
-    let metadata = snapshot_metadata(&state.status_map, &state.host_user, &state.chat_path).await;
-    let available_commands = state.plugin_registry.all_commands();
-
-    let team_access: Option<Box<dyn room_protocol::plugin::TeamAccess>> = state
-        .auth
-        .registry
-        .get()
-        .map(|reg| Box::new(crate::plugin::bridge::TeamChecker::new(reg)) as _);
-
-    CommandContext {
-        command: cmd.to_owned(),
-        params: params.to_vec(),
-        sender: username.to_owned(),
-        room_id: state.room_id.as_ref().clone(),
-        message_id: id.to_owned(),
-        timestamp: ts,
-        history: Box::new(history),
-        writer: Box::new(writer),
-        metadata,
-        available_commands,
-        team_access,
-    }
-}
-
-/// Run a plugin handler with timeout and panic isolation.
-///
-/// Wraps the handler in [`catch_unwind`] and [`tokio::time::timeout`] so that
-/// panicking or hung plugins do not take down the broker.
-async fn run_plugin_with_catch(
-    plugin: &dyn crate::plugin::Plugin,
-    ctx: CommandContext,
-) -> anyhow::Result<PluginResult> {
-    match tokio::time::timeout(
-        PLUGIN_TIMEOUT,
-        AssertUnwindSafe(plugin.handle(ctx)).catch_unwind(),
-    )
-    .await
-    {
-        Ok(Ok(r)) => r,
-        Ok(Err(panic_info)) => {
-            let msg = panic_info
-                .downcast_ref::<String>()
-                .map(|s| s.as_str())
-                .or_else(|| panic_info.downcast_ref::<&str>().copied())
-                .unwrap_or("unknown panic");
-            eprintln!("[broker] plugin '{}' panicked: {msg}", plugin.name());
-            Ok(PluginResult::Reply(format!(
-                "plugin '{}' panicked: {msg}",
-                plugin.name()
-            )))
-        }
-        Err(_elapsed) => {
-            eprintln!(
-                "[broker] plugin '{}' timed out after {}s",
-                plugin.name(),
-                PLUGIN_TIMEOUT.as_secs()
-            );
-            Ok(PluginResult::Reply(format!(
-                "plugin '{}' timed out after {}s",
-                plugin.name(),
-                PLUGIN_TIMEOUT.as_secs()
-            )))
-        }
-    }
-}
-
-/// Translate a [`PluginResult`] into a [`CommandResult`].
-///
-/// For cross-room dispatch, pass `cross_room = Some((source_room_id, target_room_id))`
-/// so that replies are tagged with the target room prefix and broadcasts go to the
-/// correct room's clients. For local dispatch, pass `None`.
-async fn translate_plugin_result(
-    result: PluginResult,
-    plugin_name: &str,
-    state: &RoomState,
-    cross_room: Option<(&str, &str)>,
-) -> anyhow::Result<CommandResult> {
-    Ok(match result {
-        PluginResult::Reply(text) => {
-            let reply_text = match cross_room {
-                Some((_, target)) => format!("[→{target}] {text}"),
-                None => text,
-            };
-            let reply_room = match cross_room {
-                Some((source, _)) => source,
-                None => &state.room_id,
-            };
-            let sys = make_system(reply_room, &format!("plugin:{plugin_name}"), reply_text);
-            let json = serde_json::to_string(&sys)?;
-            CommandResult::Reply(json)
-        }
-        PluginResult::Broadcast(text) => {
-            let sys = make_system(&state.room_id, &format!("plugin:{plugin_name}"), text);
-            let seq_msg =
-                broadcast_and_persist(&sys, &state.clients, &state.chat_path, &state.seq_counter)
-                    .await?;
-            let json = serde_json::to_string(&seq_msg)?;
-            CommandResult::HandledWithReply(json)
-        }
-        PluginResult::Handled => CommandResult::Handled,
-    })
-}
-
-/// Build a [`CommandContext`] and call a plugin's `handle` method, translating
-/// the [`PluginResult`] into a [`CommandResult`] the broker understands.
-///
-/// If params contain `--room <id>` and the source room has a cross-room
-/// resolver (daemon mode), the command is executed against the target room
-/// instead of the source room. The plugin is resolved from the target room's
-/// registry. This enables e.g. `/taskboard post --room other-room description`.
-async fn dispatch_plugin(
-    plugin: &dyn crate::plugin::Plugin,
-    msg: &Message,
-    username: &str,
-    state: &RoomState,
-) -> anyhow::Result<CommandResult> {
-    let (cmd, params, id, ts) = match msg {
-        Message::Command {
-            cmd,
-            params,
-            id,
-            ts,
-            ..
-        } => (cmd, params, id, ts),
-        _ => return Ok(CommandResult::Passthrough(msg.clone())),
-    };
-
-    // Check for --room flag to redirect to a different room.
-    if let Some((target_room_id, cleaned_params)) = extract_room_flag(params) {
-        return dispatch_cross_room(CrossRoomContext {
-            cmd,
-            params: &cleaned_params,
-            id,
-            ts: *ts,
-            username,
-            source_state: state,
-            target_room_id: &target_room_id,
-        })
-        .await;
-    }
-
-    // Schema validation — check params against the plugin's declared schema
-    // before invoking the handler.
-    if let Some(cmd_info) = plugin.commands().iter().find(|c| c.name == *cmd) {
-        if let Err(err_msg) = validate_params(params, cmd_info) {
-            let sys = make_system(
-                &state.room_id,
-                &format!("plugin:{}", plugin.name()),
-                err_msg,
-            );
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    }
-
-    let ctx = build_command_context(cmd, params, id, *ts, username, state, plugin.name()).await;
-
-    let result = run_plugin_with_catch(plugin, ctx).await?;
-
-    translate_plugin_result(result, plugin.name(), state, None).await
-}
-
-/// Bundles the parameters for cross-room plugin dispatch into a single struct,
-/// keeping the function signature readable and easy to extend.
-struct CrossRoomContext<'a> {
-    cmd: &'a str,
-    params: &'a [String],
-    id: &'a str,
-    ts: chrono::DateTime<chrono::Utc>,
-    username: &'a str,
-    source_state: &'a RoomState,
-    target_room_id: &'a str,
-}
-
-/// Execute a plugin command against a different room (cross-room dispatch).
-///
-/// Resolves the target room via `state.cross_room_resolver`, finds the plugin
-/// in the target room's registry, builds a `CommandContext` pointing at the
-/// target room's state, and runs the plugin handler. Broadcasts go to the
-/// target room's clients and chat file.
-async fn dispatch_cross_room(cx: CrossRoomContext<'_>) -> anyhow::Result<CommandResult> {
-    let CrossRoomContext {
-        cmd,
-        params,
-        id,
-        ts,
-        username,
-        source_state,
-        target_room_id,
-    } = cx;
-
-    let resolver = match source_state.cross_room_resolver.get() {
-        Some(r) => r,
-        None => {
-            let sys = make_system(
-                &source_state.room_id,
-                "broker",
-                "cross-room commands are only available in daemon mode",
-            );
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    };
-
-    let target_state = match resolver.resolve_room(target_room_id).await {
-        Some(s) => s,
-        None => {
-            let sys = make_system(
-                &source_state.room_id,
-                "broker",
-                format!("room not found: {target_room_id}"),
-            );
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    };
-
-    let plugin = match target_state.plugin_registry.resolve(cmd) {
-        Some(p) => p,
-        None => {
-            let sys = make_system(
-                &source_state.room_id,
-                "broker",
-                format!("command /{cmd} not found in room {target_room_id}"),
-            );
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    };
-
-    // Schema validation against the target plugin's schema.
-    if let Some(cmd_info) = plugin.commands().iter().find(|c| c.name == cmd) {
-        if let Err(err_msg) = validate_params(params, cmd_info) {
-            let sys = make_system(
-                &source_state.room_id,
-                &format!("plugin:{}", plugin.name()),
-                err_msg,
-            );
-            let json = serde_json::to_string(&sys)?;
-            return Ok(CommandResult::Reply(json));
-        }
-    }
-
-    // Build context against the TARGET room's state.
-    let ctx =
-        build_command_context(cmd, params, id, ts, username, &target_state, plugin.name()).await;
-
-    let result = run_plugin_with_catch(plugin, ctx).await?;
-
-    // Replies go to the SOURCE room (the caller sees them), but broadcasts go
-    // to the TARGET room (where the state change happened).
-    translate_plugin_result(
-        result,
-        plugin.name(),
-        &target_state,
-        Some((&source_state.room_id, target_room_id)),
-    )
-    .await
-}
-
-// ── Room management commands ──────────────────────────────────────────────────
-
-/// Handle `/info [username]` — show room metadata or user info.
-///
-/// - `/info` (no args): room ID, visibility, config, host, member count, subscribers.
-/// - `/info <username>`: status, subscription tier, host flag, online status.
-///
-/// `/room-info` is an alias that always shows room info (ignores params).
-async fn handle_info(params: &[String], state: &RoomState) -> String {
-    if let Some(target) = params.first() {
-        let target = target.strip_prefix('@').unwrap_or(target);
-        return handle_user_info(target, state).await;
-    }
-    handle_room_info(state).await
-}
-
-/// Room metadata: visibility, config, host, member count, subscribers.
-async fn handle_room_info(state: &RoomState) -> String {
-    let member_count = state.status_count().await;
-    let sub_count = state.subscription_count().await;
-    let host = state
-        .host_user
-        .lock()
-        .await
-        .clone()
-        .unwrap_or_else(|| "(none)".to_owned());
-
-    match &state.config {
-        Some(config) => {
-            let vis = serde_json::to_string(&config.visibility).unwrap_or_default();
-            let max = config
-                .max_members
-                .map(|n| n.to_string())
-                .unwrap_or_else(|| "unlimited".to_owned());
-            let invites: Vec<&str> = config.invite_list.iter().map(|s| s.as_str()).collect();
-            format!(
-                "room: {} | visibility: {} | max members: {} | host: {} | members online: {} | subscribers: {} | invited: [{}] | created by: {}",
-                state.room_id, vis, max, host, member_count, sub_count, invites.join(", "), config.created_by
-            )
-        }
-        None => {
-            format!(
-                "room: {} | visibility: public (legacy) | host: {} | members online: {} | subscribers: {}",
-                state.room_id, host, member_count, sub_count
-            )
-        }
-    }
-}
-
-/// User info: online status, status text, subscription tier, host flag.
-async fn handle_user_info(target: &str, state: &RoomState) -> String {
-    let status_map = state.status_map.lock().await;
-    let online = status_map.contains_key(target);
-    let status_text = status_map.get(target).cloned().unwrap_or_default();
-    drop(status_map);
-
-    let sub_tier = state
-        .filters
-        .subscription_map
-        .lock()
-        .await
-        .get(target)
-        .copied();
-
-    let is_host = state.host_user.lock().await.as_deref() == Some(target);
-
-    let mut parts = vec![format!("user: {target}")];
-
-    if online {
-        if status_text.is_empty() {
-            parts.push("online".to_owned());
-        } else {
-            parts.push(format!("online ({status_text})"));
-        }
-    } else {
-        parts.push("offline".to_owned());
-    }
-
-    if let Some(tier) = sub_tier {
-        parts.push(format!("subscription: {tier}"));
-    } else {
-        parts.push("subscription: none".to_owned());
-    }
-
-    if is_host {
-        parts.push("host: yes".to_owned());
-    }
-
-    parts.join(" | ")
-}
-
-// ── Help command ──────────────────────────────────────────────────────────────
-
-/// Handle `/help [command]` — lists all commands or shows detail for one.
-fn handle_help(params: &[String], state: &RoomState) -> String {
-    let builtins = builtin_command_infos();
-    let plugin_cmds = state.plugin_registry.all_commands();
-
-    if let Some(target) = params.first() {
-        let target = target.strip_prefix('/').unwrap_or(target);
-
-        // Check plugin commands first (matches original behaviour)
-        if let Some(cmd) = plugin_cmds.iter().find(|c| c.name == target) {
-            return format_command_help(cmd);
-        }
-        // Then builtins
-        if let Some(cmd) = builtins.iter().find(|c| c.name == target) {
-            return format_command_help(cmd);
-        }
-        return format!("unknown command: /{target}");
-    }
-
-    // List all commands: builtins first, then plugins
-    let mut lines = vec!["available commands:".to_owned()];
-    for cmd in &builtins {
-        lines.push(format!("  {} — {}", cmd.usage, cmd.description));
-    }
-    for cmd in &plugin_cmds {
-        lines.push(format!("  {} — {}", cmd.usage, cmd.description));
-    }
-    lines.join("\n")
-}
-
-/// Format detailed help for a single command, including typed parameter info.
-fn format_command_help(cmd: &CommandInfo) -> String {
-    let mut lines = vec![cmd.usage.clone(), format!("  {}", cmd.description)];
-    if !cmd.params.is_empty() {
-        lines.push("  parameters:".to_owned());
-        for p in &cmd.params {
-            let req = if p.required { "required" } else { "optional" };
-            let type_hint = match &p.param_type {
-                ParamType::Text => "text".to_owned(),
-                ParamType::Username => "username".to_owned(),
-                ParamType::Number { min, max } => match (min, max) {
-                    (Some(lo), Some(hi)) => format!("number ({lo}..{hi})"),
-                    (Some(lo), None) => format!("number ({lo}..)"),
-                    (None, Some(hi)) => format!("number (..{hi})"),
-                    (None, None) => "number".to_owned(),
-                },
-                ParamType::Choice(values) => {
-                    format!("one of: {}", values.join(", "))
-                }
-            };
-            lines.push(format!(
-                "    <{}> — {} [{}] {}",
-                p.name, p.description, req, type_hint
-            ));
-        }
-    }
-    lines.join("\n")
-}
-
-// ── Tests ─────────────────────────────────────────────────────────────────────
-
 #[cfg(test)]
 mod tests {
-    use super::{handle_info, handle_room_info, handle_user_info, route_command, CommandResult};
+    use super::super::info::{handle_info, handle_room_info, handle_user_info};
+    use super::super::{route_command, CommandResult};
     use crate::broker::{
         persistence::{
             load_event_filter_map, load_subscription_map, save_event_filter_map,
@@ -1387,7 +351,7 @@ mod tests {
     // ── validate_params tests ─────────────────────────────────────────────
 
     mod validation_tests {
-        use super::super::validate_params;
+        use super::super::super::validate::validate_params;
         use crate::plugin::{CommandInfo, ParamSchema, ParamType};
 
         fn cmd_with_params(params: Vec<ParamSchema>) -> CommandInfo {
@@ -2635,7 +1599,7 @@ mod tests {
     // ── extract_room_flag tests ──────────────────────────────────────────
 
     mod cross_room_tests {
-        use super::super::extract_room_flag;
+        use super::super::super::plugin::extract_room_flag;
 
         #[test]
         fn extract_room_flag_present_at_start() {
@@ -2841,7 +1805,8 @@ mod tests {
         let state = make_state(tmp.path().to_path_buf());
         let msg = make_command("test-room", "alice", "boom", vec![]);
 
-        let result = super::dispatch_plugin(&PanicPlugin, &msg, "alice", &state).await;
+        let result =
+            super::super::plugin::dispatch_plugin(&PanicPlugin, &msg, "alice", &state).await;
 
         // Must not propagate the panic — should return Ok with an error reply
         let result = result.expect("dispatch_plugin should not propagate panic");
@@ -2892,9 +1857,10 @@ mod tests {
         let state = make_state(tmp.path().to_path_buf());
         let msg = make_command("test-room", "alice", "kaboom", vec![]);
 
-        let result = super::dispatch_plugin(&StringPanicPlugin, &msg, "alice", &state)
-            .await
-            .expect("should not propagate panic");
+        let result =
+            super::super::plugin::dispatch_plugin(&StringPanicPlugin, &msg, "alice", &state)
+                .await
+                .expect("should not propagate panic");
 
         match result {
             CommandResult::Reply(json) => {
@@ -2939,7 +1905,7 @@ mod tests {
         let state = make_state(tmp.path().to_path_buf());
         let msg = make_command("test-room", "alice", "greet", vec![]);
 
-        let result = super::dispatch_plugin(&OkPlugin, &msg, "alice", &state)
+        let result = super::super::plugin::dispatch_plugin(&OkPlugin, &msg, "alice", &state)
             .await
             .expect("normal plugin should succeed");
 
@@ -2998,7 +1964,8 @@ mod tests {
         // machinery via tokio::time::pause.
         tokio::time::pause();
 
-        let result = super::dispatch_plugin(&SlowPlugin, &msg, "alice", &state).await;
+        let result =
+            super::super::plugin::dispatch_plugin(&SlowPlugin, &msg, "alice", &state).await;
 
         let result = result.expect("dispatch_plugin should not error on timeout");
         match result {
