+++ crates/room-daemon/src/plugin/schema.rs
@@ -1,6 +1,6 @@
 use room_protocol::plugin::{CommandInfo, ParamSchema, ParamType, Plugin};
 
-use super::{agent, queue, stats, taskboard};
+use super::{queue, stats, taskboard};
 
 // ── Built-in command schemas ───────────────────────────────────────────────
 
@@ -231,7 +231,6 @@ pub fn all_known_commands() -> Vec<CommandInfo> {
     cmds.extend(queue::QueuePlugin::default_commands());
     cmds.extend(stats::StatsPlugin.commands());
     cmds.extend(taskboard::TaskboardPlugin::default_commands());
-    cmds.extend(agent::AgentPlugin::default_commands());
     cmds
 }
